refactor: exceptions location
