+++ src/main/java/com/aurum/vmmanager/domain/exceptions/CpfNotValidException.java
@@ -1,4 +1,4 @@
-package com.aurum.vmmanager.domain.exceptions.domain;
+package com.aurum.vmmanager.domain.exceptions;
 
 public class CpfNotValidException extends Exception {
     public CpfNotValidException() {
