+++ src/main/java/com/aurum/vmmanager/domain/exceptions/NameNotValidException.java
@@ -1,4 +1,4 @@
-package com.aurum.vmmanager.domain.exceptions.domain;
+package com.aurum.vmmanager.domain.exceptions;
 
 public class NameNotValidException extends Exception {
     public NameNotValidException() {
